Relocate S3Scanner tests
