+++ test/support/src/vfs_helpers.cc
@@ -417,86 +417,4 @@ std::string TemporaryDirectoryFixture::create_temporary_array(
   return array_uri;
 }
 
-VFSTestBase::VFSTestBase(
-    const std::vector<size_t>& test_tree, const std::string& prefix)
-    : test_tree_(test_tree)
-    , compute_(4)
-    , io_(4)
-    , vfs_(&tiledb::test::g_helper_stats, &io_, &compute_, create_test_config())
-    , prefix_(prefix)
-    , temp_dir_(tiledb::test::test_dir(prefix_))
-    , is_supported_(vfs_.supports_uri_scheme(temp_dir_)) {
-  // TODO: Throw when we can provide a list of supported filesystems to Catch2.
-}
-
-VFSTestBase::~VFSTestBase() {
-  if (vfs_.supports_uri_scheme(temp_dir_)) {
-    bool is_dir = false;
-    vfs_.is_dir(temp_dir_, &is_dir).ok();
-    if (is_dir) {
-      vfs_.remove_dir(temp_dir_).ok();
-    }
-  }
-}
-
-tiledb::sm::Config VFSTestBase::create_test_config() {
-  tiledb::sm::Config cfg;
-  if constexpr (!tiledb::test::aws_s3_config) {
-    // Set up connection to minio backend emulator.
-    cfg.set("vfs.s3.endpoint_override", "localhost:9999").ok();
-    cfg.set("vfs.s3.scheme", "https").ok();
-    cfg.set("vfs.s3.use_virtual_addressing", "false").ok();
-    cfg.set("vfs.s3.verify_ssl", "false").ok();
-  }
-  cfg.set("vfs.azure.storage_account_name", "devstoreaccount1").ok();
-  cfg.set(
-         "vfs.azure.storage_account_key",
-         "Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/"
-         "K1SZFPTOtr/KBHBeksoGMGw==")
-      .ok();
-  cfg.set("vfs.azure.blob_endpoint", "http://127.0.0.1:10000/devstoreaccount1")
-      .ok();
-  return cfg;
-}
-
-VFSTest::VFSTest(
-    const std::vector<size_t>& test_tree, const std::string& prefix)
-    : VFSTestBase(test_tree, prefix) {
-  if (!is_supported()) {
-    return;
-  }
-
-  if (temp_dir_.is_file() || temp_dir_.is_memfs() || temp_dir_.is_hdfs()) {
-    vfs_.create_dir(temp_dir_).ok();
-  } else {
-    vfs_.create_bucket(temp_dir_).ok();
-  }
-  for (size_t i = 1; i <= test_tree_.size(); i++) {
-    sm::URI path = temp_dir_.join_path("subdir_" + std::to_string(i));
-    // VFS::create_dir is a no-op for S3.
-    vfs_.create_dir(path).ok();
-    for (size_t j = 1; j <= test_tree_[i - 1]; j++) {
-      auto object_uri = path.join_path("test_file_" + std::to_string(j));
-      vfs_.touch(object_uri).ok();
-      std::string data(j * 10, 'a');
-      vfs_.open_file(object_uri, sm::VFSMode::VFS_WRITE).ok();
-      vfs_.write(object_uri, data.data(), data.size()).ok();
-      vfs_.close_file(object_uri).ok();
-      expected_results().emplace_back(object_uri.to_string(), data.size());
-    }
-  }
-  std::sort(expected_results().begin(), expected_results().end());
-}
-
-LocalFsTest::LocalFsTest(const std::vector<size_t>& test_tree)
-    : VFSTestBase(test_tree, "file://") {
-#ifdef _WIN32
-  temp_dir_ =
-      tiledb::test::test_dir(prefix_ + tiledb::sm::Win::current_dir() + "/");
-#else
-  temp_dir_ =
-      tiledb::test::test_dir(prefix_ + tiledb::sm::Posix::current_dir() + "/");
-#endif
-}
-
 }  // namespace tiledb::test
