+++ tiledb/sm/filesystem/test/CMakeLists.txt
@@ -26,8 +26,9 @@
 include(unit_test)
 
 commence(unit_test vfs)
-    this_target_object_libraries(vfs)
     this_target_sources(main.cc unit_uri.cc unit_ls_filtered.cc)
+    # Object store definitions and libraries needed for cloud backends.
+    this_target_link_libraries(TILEDB_CORE_OBJECTS TILEDB_CORE_OBJECTS_ILIB)
 conclude(unit_test)
 
 commence(unit_test vfs_read_log_modes)
